requirements.txt
pdfplumber>=0.10.0
openpyxl>=3.1.0
pydantic>=2.5.0
requests>=2.31.0
PyYAML>=6.0
rich>=13.7.0
click>=8.1.0
